+++ application/usecases/auth/dto/DecodedAccessToken.ts
@@ -1,0 +1,4 @@
+export interface DecodedAccessToken {
+  userId: string;
+  exp: number;
+}
